update ideas, add send idea on market (50%)

+++ HITSBlazor/Utils/Models/Formulas.cs
@@ -3,6 +3,6 @@
     public static class Formulas
     {
         public static double CalculcateRating(params int[] values)
-            => Math.Round((double)values.Sum() / (double)values.Length, 1);
+            => Math.Round((double)values.Sum() / values.Length, 1);
     }
 }
